Integrate pul_metadata_services into remote_record
