--- app/services/pul_metadata_services.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-module PulMetadataServices
-end
